refactor(cli): split layout into api/layout leaf shape
Reshape the flat api/layout/layout.mjs into the fractal leaf shape: a shared
_adapter.mjs (analyze + loadBlocks + formatIssue over lib/xle, used by
expand/check) with thin expand/, check/, and grammar/ leaves, plus a layout.mjs
barrel. api/index.mjs, the CLI consumer, and layout.test.mjs are unchanged —
they all import via the barrel, so the functions just relocate.

Pure reorg: layout expand / check / grammar --json envelopes and human output
are byte-identical (verified old-vs-new differential across 9 cases).
